+++ src/accessories/childlock/childlock.handler.js
@@ -2,8 +2,6 @@
 
 const logger = require('../../utils/logger');
 
-const timeout = (ms) => new Promise((res) => setTimeout(res, ms));
-
 class Handler {
   constructor() {
     this.configured = false;
@@ -21,7 +19,7 @@ class Handler {
 
     this.configured = true;
 
-    this.poll();
+    setTimeout(() => this.poll(), 1000);
 
     return this;
   }
@@ -96,8 +94,6 @@ class Handler {
   }
 
   async poll() {
-    await timeout(1000); //wait for accessories to fully load
-
     try {
       const accessories = this.accessories.filter((accessory) => accessory.context.config.type === 'childlock');