+++ src/accessories/extras/extras.handler.js
@@ -6,8 +6,6 @@ const { requestLUA, requestXml } = require('@seydx/fritzbox/lib/request');
 const logger = require('../../utils/logger');
 const Telegram = require('../../lib/telegram');
 
-const timeout = (ms) => new Promise((res) => setTimeout(res, ms));
-
 class Handler {
   constructor() {
     this.alarmTimeout = null;
@@ -29,7 +27,7 @@ class Handler {
 
     this.configured = true;
 
-    this.poll();
+    setTimeout(() => this.poll(), 1000);
 
     return this;
   }
@@ -661,7 +659,6 @@ class Handler {
   }
 
   async poll() {
-    await timeout(1000); //wait for accessories to fully load
     const RouterHandler = require('../router/router.handler');
 
     const validOptionsSwitches = [