+++ src/accessories/presence/presence.handler.js
@@ -6,8 +6,6 @@ const logger = require('../../utils/logger');
 const { validIP } = require('../../utils/utils');
 const Telegram = require('../../lib/telegram');
 
-const timeout = (ms) => new Promise((res) => setTimeout(res, ms));
-
 class Handler {
   constructor() {
     this.hosts = [];
@@ -26,7 +24,7 @@ class Handler {
 
     this.configured = true;
 
-    this.poll();
+    setTimeout(() => this.poll(), 1000);
 
     return this;
   }
@@ -232,8 +230,6 @@ class Handler {
   async set(state, accessory, subtype, ownCharacteristic) {}
 
   async poll() {
-    await timeout(1000); //wait for accessories to fully load
-
     try {
       this.hosts = await this.fritzbox.getAllHostsV2();
       //logger.debug(this.hosts, 'Hosts');